Differentiate each metric by kind
On the metrics page, unless you look at the axis labels and units, it is
not obvious what type of metric you're looking at. This manifests when
looking at metrics for the same event. The same title will be duplicated
each time without displaying the metric.

This PR is mostly to get a conversation started. I'm not sure of the
best way to display this information. For now I displayed is as a
function acting on the data.

Separately, it also may be nice to display the `event_name` rather than
`name`, and highlight the measurement differently.