More robust metrics history example (#201)
In the MyApp.MyStorage example in the metrics history guide, the
`attach_handler` function should rely on the :event_name key rather
than the :name key in the metric. The :event_name should always be the
full event name to attach to, whereas :name might be different if the
metric was configured with separate name, :event_name, and :measurement
values (to give the metric a different label in the dashboard).