Fix Ecto stats page when running on multi node env (#327)
It fixes the auto discovery feature of Ecto repos when running in a
distribution environment.
This is specially important when the nodes connected are heterogeneous.

The approach is optimistic after detection of running repos. This means
that RPC errors are not being treated after running repos detection,
which is fine in most cases.

This change also includes an update to the `ecto_psql_extras` package,
which now supports sending queries to remote nodes.